update column type conapo metro area

+++ etl/metro_areas_population/dim_metro_areas_population.py
@@ -35,7 +35,7 @@ class MetroAreaPopulationPipeline(EasyPipeline):
 
         dtype = {
             "zm_id":   "UInt32",
-            "mun_id":  "UInt32",
+            "mun_id":  "UInt16",
         }
 
         download_step = DownloadStep(
